refactor: move upload registration helpers to executor_files_upload_register.go
